+++ CHANGES.rst
@@ -12,6 +12,8 @@ Version 0.9.1 (in development)
 
 * Added
 
+  - the `server32_dir` keyword argument to :class:`~msl.loadlib.client64.Client64`
+    (fixes issue `#35 <https://github.com/MSLNZ/msl-loadlib/issues/35>`_)
   - Support for Python 3.11
   - Support for Python 3.10
   - :class:`~msl.loadlib.load_library.LoadLibrary` and :class:`~msl.loadlib.client64.Client64`